Completed: SCRUM-297 (Page Layout)), SCRUM-309(Redirect to other Pages), SCRUM-310(Edit Feature), and SCRUM-311('+' and '-' buttons))
Completed the following SCRUM tasks:
SCRUM-297: Created the general layout of the page as can be seen on FIGMA/Jira.
SCRUM-309: Created a sidebar menu for the other admin pages. Added functionality to each button to redirect user to the desired page.
SCRUM-310: Added an 'edit' button to each item to change the value, description, and name of each item in the list.
SCRUM-311: Added '+' and '-' buttons to each item in list to quickly add or subtract from each item.