#375 Images are saved into the S3 Buckets
I was able to make a new AWS account that will hold the s3 bucket for this class. I made IAM group so everyone can get credentials to make the S3 bucket work.

On the Online-estimates page, if a user uploads the images and click confirm, the images they uploaded will then be saved into the S3 bucket.

Setup instructions are in the wiki\setup.md